+++ src/Cwaff-Guns/Zag.cs
@@ -14,11 +14,13 @@ public class Zag : CwaffGun
     {
         Gun gun = Lazy.SetupGun<Zag>(ItemName, ShortDescription, LongDescription, Lore);
             gun.SetAttributes(quality: ItemQuality.A, gunClass: GunClass.PISTOL, reloadTime: 0.8f, ammo: 400, shootFps: 30, reloadFps: 40,
-                fireAudio: "zag_zig_sound", reloadAudio: "zag_zig_sound");
+                fireAudio: "zag_zig_sound", reloadAudio: "zag_zig_sound", muzzleFrom: Items.Heroine);
             gun.LoopAnimation(gun.reloadAnimation);
 
         gun.InitProjectile(GunData.New(clipSize: 9, cooldown: 0.125f, shootStyle: ShootStyle.SemiAutomatic,
-          damage: 5.0f, speed: 40.0f, sprite: "zag_bullet", fps: 8, anchor: Anchor.MiddleCenter)).Attach<ZagProjectile>();
+            damage: 5.0f, speed: 40.0f, sprite: "zag_bullet", fps: 8, anchor: Anchor.MiddleCenter, hitEnemySound: "zag_hit_enemy_sound"))
+          .Attach<ZagProjectile>()
+          .CopyAllImpactVFX(Items.ShockRifle);
 
         _ZagTrailPrefab = VFX.CreateTrailObject("zag_trail_mid", fps: 30, cascadeTimer: C.FRAME, destroyOnEmpty: true);